sdk: PPU headers compile standalone under -Wall -Wextra -Werror
Every installed PPU header is now compiled on its own in C (gnu99, c11)
and C++17, on both ABIs, under -Wall -Wextra -Werror.  Fixes from that
probe:

- ppu-asm.h: drop 'register' from the statement-expression temporaries,
  which C++17 rejects.
- Syscall wrappers widen pointers through uintptr_t instead of a direct
  pointer-to-u64 cast, which is a different-size cast on ILP32
  (sys/lv2_fs_ext.h, sys/process.h, sys/synchronization.h, sys/thread.h,
  sys/ppu_thread.h).
- The SDK now owns sys/mutex.h, sys/cond.h, sys/event_queue.h, sys/sem.h,
  sys/systime.h and sys/spu.h with the same widening, and PSL1GHT is
  installed by scripts/install-psl1ght.sh so a PSL1GHT reinstall no
  longer overwrites SDK-owned headers.
- cell/dbgfont.h includes <stdarg.h>; cell/sail/player.h includes the
  descriptor header it uses.
- Vectormath: Vector3, Vector4, Point3, Quat, floatInVec and boolInVec
  lose their member-wise copy assignments, whose presence made the
  implicit copy constructor deprecated.  Copying is unchanged: the copy
  constructors were and remain trivial, so passing by value is the same,
  and the types are now also trivially copyable.  Source that names
  the removed operator (taking &T::operator=) no longer compiles;
  ordinary assignment is unaffected.

tests/sdk/ppu-header-hygiene-test.py is the probe.  Headers that are
fragments of another header are classified with their owner; known
failures that remain are listed with their exact diagnostics in
ppu-header-hygiene-residuals.json, and the test fails when a listed
failure changes or is fixed without being removed, or when a new one
appears.  tests/sdk/psl1ght-header-install-order-test.py checks the
install ownership.

+++ sdk/include/sys/lv2_fs_ext.h
@@ -25,6 +25,7 @@
 #ifndef __PS3DK_SYS_LV2_FS_EXT_H__
 #define __PS3DK_SYS_LV2_FS_EXT_H__
 
+#include <stdint.h>
 #include <ppu-types.h>
 #include <sys/lv2_syscall.h>
 
@@ -36,7 +37,7 @@ extern "C" {
 LV2_SYSCALL
 sysLv2FsChown(const char *path, s32 uid, s32 gid)
 {
-	lv2syscall3(835, (u64)path, (u64)uid, (u64)gid);
+	lv2syscall3(835, (u64)(uintptr_t)path, (u64)uid, (u64)gid);
 	return_to_user_prog(s32);
 }
 
@@ -51,8 +52,8 @@ LV2_SYSCALL
 sysLv2FsMount(const char *deviceName, const char *deviceFileSystem,
 	      const char *devicePath, s32 writeProt)
 {
-	lv2syscall8(837, (u64)deviceName, (u64)deviceFileSystem,
-		    (u64)devicePath, 0, (u64)writeProt, 0, 0, 0);
+	lv2syscall8(837, (u64)(uintptr_t)deviceName, (u64)(uintptr_t)deviceFileSystem,
+		    (u64)(uintptr_t)devicePath, 0, (u64)writeProt, 0, 0, 0);
 	return_to_user_prog(s32);
 }
 
@@ -60,7 +61,7 @@ sysLv2FsMount(const char *deviceName, const char *deviceFileSystem,
 LV2_SYSCALL
 sysLv2FsUnmount(const char *devicePath)
 {
-	lv2syscall3(838, (u64)devicePath, 0, 0);
+	lv2syscall3(838, (u64)(uintptr_t)devicePath, 0, 0);
 	return_to_user_prog(s32);
 }
